+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 0.3.0
+* added keepalive node
+* correct handling of online-node with more than one miniserver connection
+
+### 0.2.2 
+* fixed error if empty room name or empty category name
+
 ### 0.2.1
 * fixed error on deploying after editing the config node
 
@@ -9,7 +16,8 @@
 
 ### 0.1.0
 * breaks compatibility to 0.0.x
-* updated `node-lox-ws`
+* added webservice-node
+* updated `node-lox-ws-api`
 
 ### 0.0.x
 * control-in